集成:百度uid(修改用户实体主键为uid)

+++ common/src/main/java/com/huimu/flowerplant/common/model/BaseEntity.java
@@ -11,9 +11,9 @@ import java.util.Date;
 @ToString
 public class BaseEntity {
 
-    private Integer createdUserId;
+    private Long createdUserId;
 
-    private Integer updatedUserId;
+    private Long updatedUserId;
 
     private Date createdTime;
